+++ lib/view_component.rb
@@ -12,4 +12,5 @@ module ViewComponent
   autoload :TestHelpers
   autoload :TestCase
   autoload :TemplateError
+  autoload :SidecarI18n
 end