+++ test/app/components/sidecar_i18n_component.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class SidecarI18nComponent < ViewComponent::Base
+  include ViewComponent::SidecarI18n
+end